+++ dtfs-central-api/src/v1/controllers/utilisation-report-service/get-utilisation-reports-reconciliation-summary.controller/helpers/reconciliation-summary-generator.test.ts
@@ -8,12 +8,12 @@ import {
   IsoMonthStamp,
 } from '@ukef/dtfs2-common';
 
-import { UtilisationReportRepo } from '../../../../repositories/utilisation-reports-repo';
-import { generateReconciliationSummaries, getAllReportsForSubmissionMonth, getPreviousOpenReportsBySubmissionMonth } from './helpers';
-import { UtilisationReportReconciliationSummary, UtilisationReportReconciliationSummaryItem } from '../../../../types/utilisation-reports';
-import { aBank } from '../../../../../test-helpers/test-data/bank';
-import { aMonthlyBankReportPeriodSchedule } from '../../../../../test-helpers/test-data/bank-report-period-schedule';
-import { getAllBanks } from '../../../../repositories/banks-repo';
+import { UtilisationReportRepo } from '../../../../../repositories/utilisation-reports-repo';
+import { generateReconciliationSummaries, getAllReportsForSubmissionMonth, getPreviousOpenReportsBySubmissionMonth } from './reconciliation-summary-generator';
+import { UtilisationReportReconciliationSummary, UtilisationReportReconciliationSummaryItem } from '../../../../../types/utilisation-reports';
+import { aBank } from '../../../../../../test-helpers/test-data/bank';
+import { aMonthlyBankReportPeriodSchedule } from '../../../../../../test-helpers/test-data/bank-report-period-schedule';
+import { getAllBanks } from '../../../../../repositories/banks-repo';
 
 jest.mock('../../../../repositories/banks-repo');
 jest.mock('../../../../repositories/utilisation-reports-repo');
