+++ dtfs-central-api/src/v1/controllers/utilisation-report-service/get-utilisation-reports-reconciliation-summary.controller/helpers/reconciliation-summary-generator.ts
@@ -10,9 +10,9 @@ import {
   UtilisationReportEntity,
   IsoMonthStamp,
 } from '@ukef/dtfs2-common';
-import { UtilisationReportReconciliationSummary, UtilisationReportReconciliationSummaryItem } from '../../../../types/utilisation-reports';
-import { UtilisationReportRepo } from '../../../../repositories/utilisation-reports-repo';
-import { getAllBanks } from '../../../../repositories/banks-repo';
+import { UtilisationReportReconciliationSummary, UtilisationReportReconciliationSummaryItem } from '../../../../../types/utilisation-reports';
+import { UtilisationReportRepo } from '../../../../../repositories/utilisation-reports-repo';
+import { getAllBanks } from '../../../../../repositories/banks-repo';
 import { mapReportToSummaryItem } from './reconciliation-summary-item-mapper';
 
 type UtilisationReportForSubmissionMonth = {
