+++ dtfs-central-api/src/v1/controllers/utilisation-report-service/get-utilisation-reports-reconciliation-summary.controller/helpers/reconciliation-summary-item-mapper.test.ts
@@ -8,7 +8,7 @@ import {
 } from '@ukef/dtfs2-common';
 import { difference } from 'lodash';
 import { mapReportToSummaryItem } from './reconciliation-summary-item-mapper';
-import { aBank } from '../../../../../test-helpers/test-data';
+import { aBank } from '../../../../../../test-helpers/test-data';
 
 describe('reconciliation-summary-item-mapper', () => {
   describe('mapReportToSummaryItem', () => {
