+++ dtfs-central-api/src/v1/controllers/utilisation-report-service/get-utilisation-reports-reconciliation-summary.controller/helpers/reconciliation-summary-item-mapper.ts
@@ -1,5 +1,5 @@
 import { Bank, FeeRecordEntity, UtilisationReportEntity } from '@ukef/dtfs2-common';
-import { UtilisationReportReconciliationSummaryItem } from '../../../../types/utilisation-reports';
+import { UtilisationReportReconciliationSummaryItem } from '../../../../../types/utilisation-reports';
 
 const getCountOfReconciledFeeRecords = (feeRecords: FeeRecordEntity[]): number => {
   return feeRecords.filter((feeRecord) => feeRecord.status === 'RECONCILED').length;
